FEAT(PeroidPlanEditScreen): 기간으로 계획하기 계획 수정 완료

+++ domain/src/main/java/com/capston/domain/response/plan/PlanDetailResponse.kt
@@ -10,7 +10,7 @@ data class PlanDetailResponse (
     val startDate: String = "",
     val endDate: String = "",
     val dailyTime: Int = 0,
-    val playbackSpeed: Int = 0,
+    val playbackSpeed: Double = 0.0,
     val dailySchedules: List<PlanDetailDailySchedule> = emptyList(),
 )
